Move build_archive() from test_sdist to common helpers module
This will allow to reuse it in later commits. Change the function to
return a pathlib.Path instead of a string because that is what is used
internally and because it is easy to turn that into a string when
needed, but it is a bit more cumbersome to do the opposite.
