+++ tests/helpers.py
@@ -13,8 +13,12 @@
 # limitations under the License.
 """Test functions useful across twine's tests."""
 
+import io
 import os
 import pathlib
+import tarfile
+import textwrap
+import zipfile
 
 TESTS_DIR = pathlib.Path(__file__).parent
 FIXTURES_DIR = os.path.join(TESTS_DIR, "fixtures")
@@ -22,3 +26,28 @@ SDIST_FIXTURE = os.path.join(FIXTURES_DIR, "twine-1.5.0.tar.gz")
 WHEEL_FIXTURE = os.path.join(FIXTURES_DIR, "twine-1.5.0-py2.py3-none-any.whl")
 NEW_SDIST_FIXTURE = os.path.join(FIXTURES_DIR, "twine-1.6.5.tar.gz")
 NEW_WHEEL_FIXTURE = os.path.join(FIXTURES_DIR, "twine-1.6.5-py2.py3-none-any.whl")
+
+
+def build_archive(path, name, archive_format, files):
+    filepath = path / f"{name}.{archive_format}"
+
+    if archive_format == "tar.gz":
+        with tarfile.open(filepath, "x:gz") as archive:
+            for mname, content in files.items():
+                if isinstance(content, tarfile.TarInfo):
+                    content.name = mname
+                    archive.addfile(content)
+                else:
+                    data = textwrap.dedent(content).encode("utf8")
+                    member = tarfile.TarInfo(mname)
+                    member.size = len(data)
+                    archive.addfile(member, io.BytesIO(data))
+        return filepath
+
+    if archive_format == "zip":
+        with zipfile.ZipFile(filepath, mode="w") as archive:
+            for mname, content in files.items():
+                archive.writestr(mname, textwrap.dedent(content))
+        return filepath
+
+    raise ValueError(format)
